Centralize Tour components and functions
